% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pandocfilters.R
\name{astrapply}
\alias{astrapply}
\title{Apply a Function on a AST}
\usage{
astrapply(x, FUN, ...)
}
\arguments{
\item{x}{a list representing the AST obtained from pandoc.}

\item{FUN}{the function to be applied to the AST.}

\item{...}{optional arguments to \code{FUN}.}
}
\value{
A list containing the modified AST.
}
\description{
Apply the function \code{FUN} on the abstract syntax tree (AST) obtained from pandoc.
}

