% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pandocfilters.R
\name{Attr}
\alias{Attr}
\title{Attributes}
\usage{
Attr(identifier = "", classes = character(), key_val_pairs = list())
}
\arguments{
\item{identifier}{a character string}

\item{classes}{a character giving the classes}

\item{key_val_pairs}{a list of tuple of type \code{"character"}}
}
\description{
A constructor for pandoc attributes.
}
\examples{
Attr("A", c("B", "C"), list(c("D", "E")))
}

