% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/block_elements.R
\name{BlockQuote}
\alias{BlockQuote}
\title{Block Quote.}
\usage{
BlockQuote(blocks)
}
\arguments{
\item{blocks}{a block object or list of block objects}
}
\description{
Constructs a block object of type \code{BlockQuote}.
}
\examples{
BlockQuote(Plain("Hello R!"))

}
\seealso{
Other Block element constructors: \code{\link{CodeBlock}},
  \code{\link{DefinitionList}}, \code{\link{Definition}},
  \code{\link{Div}}, \code{\link{Header}},
  \code{\link{HorizontalRule}}, \code{\link{Null}},
  \code{\link{OrderedList}}, \code{\link{Para}},
  \code{\link{Plain}}, \code{\link{Table}}
}
