% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pandocfilters.R
\name{BulletList}
\alias{BulletList}
\title{Bullet List}
\usage{
BulletList(llblocks)
}
\arguments{
\item{llblocks}{a list of lists of blocks}
}
\description{
A constructor of a block object of type \code{"BulletList"}.
}
\examples{
bullet_1 <- Plain("A")
bullet_2 <- Plain(Str("B"))
bullet_3 <- list(Plain(list(Str("C"))))
BulletList(list(bullet_1, bullet_2, bullet_3))
}

