% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pandocfilters.R
\name{c.block}
\alias{c.block}
\title{Combine Block Objects}
\usage{
\method{c}{block}(...)
}
\arguments{
\item{...}{objects to be concatenated.}
}
\value{
an list of \code{"block"} objects.
}
\description{
Objects of class \code{"block"} can be combined by using the generic 
   default method \code{"c"} (combine).
}
\examples{
c(Header( "R Basics" ), Header("What is R?", level=2),
Plain(c(Emph("R"), Space(), "is a system for ", Strong("statistical computation"))))
}

