% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pandocfilters.R
\name{c.inline}
\alias{c.inline}
\title{Combine Inline Objects}
\usage{
\method{c}{inline}(...)
}
\arguments{
\item{...}{objects to be concatenated.}
}
\value{
an list of \code{"inline"} objects.
}
\description{
Objects of class \code{"inline"} can be combined by using the generic 
   default method \code{"c"} (combine).
}
\examples{
c(Str("some"), Strong("text"))
}

