% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pandocfilters.R
\name{Citation}
\alias{Citation}
\title{Citation}
\usage{
Citation(suffix, id, note_num = 0L, mode = "AuthorInText",
  prefix = list(), hash = 0L)
}
\arguments{
\item{suffix}{a inline object or list of inline objects}

\item{id}{a character string (not visible in the text)}

\item{note_num}{an integer}

\item{mode}{a character string giving the citation mode, possible values are 
\code{"AuthorInText"}, \code{"SuppressAuthor"} and \code{"NormalCitation"}.}

\item{prefix}{a inline object or list of inline objects}

\item{hash}{an integer}
}
\description{
A constructor of an object of type \code{"Citation"}.
}

