% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inline_elements.R
\name{Cite}
\alias{Cite}
\title{Citation
A constructor of an inline object of type \code{"Cite"}.}
\usage{
Cite(citation, x)
}
\arguments{
\item{citation}{an object of type \code{"Citation"}}

\item{x}{a inline object or a list of inline objects}
}
\description{
Citation
A constructor of an inline object of type \code{"Cite"}.
}
\examples{
ci <- Citation(suffix=list(Str("Suffix_1")),
               id="Citation_ID_1", prefix=list(Str("Prefix_1")))
Cite(ci, Str("some text"))

}
\seealso{
Other Inline element constructors: \code{\link{Code}},
  \code{\link{Emph}}, \code{\link{Image}},
  \code{\link{LineBreak}}, \code{\link{Link}},
  \code{\link{Math}}, \code{\link{Note}},
  \code{\link{Quoted}}, \code{\link{RawInline}},
  \code{\link{SmallCaps}}, \code{\link{SoftBreak}},
  \code{\link{Space}}, \code{\link{Span}},
  \code{\link{Strikeout}}, \code{\link{Strong}},
  \code{\link{Str}}, \code{\link{Subscript}},
  \code{\link{Superscript}}
}
