% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inline_elements.R
\name{Code}
\alias{Code}
\title{Inline Code.}
\usage{
Code(code, name = "", language = NULL, line_numbers = FALSE,
  start_from = 1)
}
\arguments{
\item{code}{a character string giving the inline code}

\item{name}{an optional character string giving the name of the inline code chunk}

\item{language}{an optional character string giving the programming language}

\item{line_numbers}{a logical which controls if line numbers should be used}

\item{start_from}{an integer giving the first line number}
}
\description{
A constructor of an inline object of type \code{"Code"}.
}
\examples{
Code("lm(hello ~ world)", "my_r_inline_code", "R", TRUE, 0)
Code("lm(hello ~ world)")

}
\seealso{
Other Inline element constructors: \code{\link{Cite}},
  \code{\link{Emph}}, \code{\link{Image}},
  \code{\link{LineBreak}}, \code{\link{Link}},
  \code{\link{Math}}, \code{\link{Note}},
  \code{\link{Quoted}}, \code{\link{RawInline}},
  \code{\link{SmallCaps}}, \code{\link{SoftBreak}},
  \code{\link{Space}}, \code{\link{Span}},
  \code{\link{Strikeout}}, \code{\link{Strong}},
  \code{\link{Str}}, \code{\link{Subscript}},
  \code{\link{Superscript}}
}
