% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/block_elements.R
\name{Definition}
\alias{Definition}
\title{Definition.}
\usage{
Definition(key, value)
}
\arguments{
\item{key}{a inline object or list of inline objects}

\item{value}{a block object or list of block objects}
}
\description{
A constructor of a \code{Definition} which can be used as an element of a \link{DefinitionList}.
}
\examples{
Definition("some key", Plain("some value"))

}
\seealso{
Other Block element constructors: \code{\link{BlockQuote}},
  \code{\link{CodeBlock}}, \code{\link{DefinitionList}},
  \code{\link{Div}}, \code{\link{Header}},
  \code{\link{HorizontalRule}}, \code{\link{Null}},
  \code{\link{OrderedList}}, \code{\link{Para}},
  \code{\link{Plain}}, \code{\link{Table}}
}
