% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/block_elements.R
\name{Div}
\alias{Div}
\title{Div is a Generic Block Container with Attributes.}
\usage{
Div(blocks, attr = Attr())
}
\arguments{
\item{blocks}{a block object or list of block objects}

\item{attr}{an object of type \code{"Attr"}}
}
\description{
A constructor of a block object of type \code{"Div"}.
}
\examples{
blocks <- Plain("Hello R!")
Div(blocks)

}
\seealso{
Other Block element constructors: \code{\link{BlockQuote}},
  \code{\link{CodeBlock}}, \code{\link{DefinitionList}},
  \code{\link{Definition}}, \code{\link{Header}},
  \code{\link{HorizontalRule}}, \code{\link{Null}},
  \code{\link{OrderedList}}, \code{\link{Para}},
  \code{\link{Plain}}, \code{\link{Table}}
}
