% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inline_elements.R
\name{Emph}
\alias{Emph}
\title{Emphasized Text.}
\usage{
Emph(x)
}
\arguments{
\item{x}{a inline object or a list of inline objects}
}
\description{
A constructor of an inline object of type \code{"Emph"}.
}
\examples{
Emph("emphasize")

}
\seealso{
Other Inline element constructors: \code{\link{Cite}},
  \code{\link{Code}}, \code{\link{Image}},
  \code{\link{LineBreak}}, \code{\link{Link}},
  \code{\link{Math}}, \code{\link{Note}},
  \code{\link{Quoted}}, \code{\link{RawInline}},
  \code{\link{SmallCaps}}, \code{\link{SoftBreak}},
  \code{\link{Space}}, \code{\link{Span}},
  \code{\link{Strikeout}}, \code{\link{Strong}},
  \code{\link{Str}}, \code{\link{Subscript}},
  \code{\link{Superscript}}
}
