% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pandocfilters.R
\name{filter}
\alias{filter}
\title{Filter JSON-formatted AST}
\usage{
filter(FUN, ..., input = stdin(), output = stdout())
}
\arguments{
\item{FUN}{the function to be applied on the AST.}

\item{...}{optional arguments to \code{FUN}.}

\item{input}{a connection object or a character string from which the JSON-formatted AST is read}

\item{output}{a connection object or a character string to which the JSON-formatted AST is written}
}
\description{
Apply a filter on the JSON-formatted abstract syntax tree (AST).
}

