% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{get_pandoc_version}
\alias{get_pandoc_version}
\title{Get Pandoc Version}
\usage{
get_pandoc_version(type = c("numeric", "character"))
}
\arguments{
\item{type}{a character giving the type of the return value.}
}
\description{
Get the version of pandoc.
}
\examples{
get_pandoc_version()
}
