% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/block_elements.R
\name{Header}
\alias{Header}
\title{Header.}
\usage{
Header(x, level = 1L, attr = Attr())
}
\arguments{
\item{x}{a inline object or a list of inline objects}

\item{level}{an integer giving the level}

\item{attr}{an object of type \link{Attr}}
}
\description{
A constructor of a block object of type \code{"Header"}.
}
\examples{
Header("My Header")

}
\seealso{
Other Block element constructors: \code{\link{BlockQuote}},
  \code{\link{CodeBlock}}, \code{\link{DefinitionList}},
  \code{\link{Definition}}, \code{\link{Div}},
  \code{\link{HorizontalRule}}, \code{\link{Null}},
  \code{\link{OrderedList}}, \code{\link{Para}},
  \code{\link{Plain}}, \code{\link{Table}}
}
