% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/block_elements.R
\name{HorizontalRule}
\alias{HorizontalRule}
\title{Horizontal Rule.}
\usage{
HorizontalRule()
}
\description{
Constructs a block object of type \code{HorizontalRule}.
}
\examples{
HorizontalRule()

}
\seealso{
Other Block element constructors: \code{\link{BlockQuote}},
  \code{\link{CodeBlock}}, \code{\link{DefinitionList}},
  \code{\link{Definition}}, \code{\link{Div}},
  \code{\link{Header}}, \code{\link{Null}},
  \code{\link{OrderedList}}, \code{\link{Para}},
  \code{\link{Plain}}, \code{\link{Table}}
}
