% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pandocfilters.R
\name{is.block}
\alias{is.block}
\title{Block Objects}
\usage{
is.block(x)
}
\arguments{
\item{x}{an object to be tested.}
}
\value{
a logical indicating if the provided object is of type \code{"block"}.
}
\description{
Tests if an object has the class attribute \code{"block"}.
}
\examples{
is.block(as.block(NULL))
}

