% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pandocfilters.R
\name{is.inline}
\alias{is.inline}
\title{Inline Objects}
\usage{
is.inline(x)
}
\arguments{
\item{x}{an object to be tested.}
}
\value{
a logical indicating if the provided object is of type \code{"inline"}.
}
\description{
Tests if an object has the class attribute \code{"inline"}.
}
\examples{
is.inline(as.inline(NULL))
}

