% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inline_elements.R
\name{LineBreak}
\alias{LineBreak}
\title{Hard Line Break.}
\usage{
LineBreak()
}
\description{
A constructor of an inline object of type \code{"LineBreak"}.
}
\examples{
LineBreak()

}
\seealso{
Other Inline element constructors: \code{\link{Cite}},
  \code{\link{Code}}, \code{\link{Emph}},
  \code{\link{Image}}, \code{\link{Link}},
  \code{\link{Math}}, \code{\link{Note}},
  \code{\link{Quoted}}, \code{\link{RawInline}},
  \code{\link{SmallCaps}}, \code{\link{SoftBreak}},
  \code{\link{Space}}, \code{\link{Span}},
  \code{\link{Strikeout}}, \code{\link{Strong}},
  \code{\link{Str}}, \code{\link{Subscript}},
  \code{\link{Superscript}}
}
