% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inline_elements.R
\name{Link}
\alias{Link}
\title{Hyperlink.}
\usage{
Link(target, text, title = "", attr = Attr())
}
\arguments{
\item{target}{a character string giving the target (hyper reference)}

\item{text}{a inline object or a list of inline objects giving the visible part}

\item{title}{an optional character string giving the title}

\item{attr}{an optional object of type \code{"Attr"}}
}
\description{
A constructor of an inline object of type \code{"Link"}.
}
\details{
Further Usage examples can be found in the README.
}
\examples{
Link("https://cran.r-project.org/", "Text_Shown", "some title")

}
\seealso{
Other Inline element constructors: \code{\link{Cite}},
  \code{\link{Code}}, \code{\link{Emph}},
  \code{\link{Image}}, \code{\link{LineBreak}},
  \code{\link{Math}}, \code{\link{Note}},
  \code{\link{Quoted}}, \code{\link{RawInline}},
  \code{\link{SmallCaps}}, \code{\link{SoftBreak}},
  \code{\link{Space}}, \code{\link{Span}},
  \code{\link{Strikeout}}, \code{\link{Strong}},
  \code{\link{Str}}, \code{\link{Subscript}},
  \code{\link{Superscript}}
}
