% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pandocfilters.R
\name{ListAttributes}
\alias{ListAttributes}
\title{ListAttributes}
\usage{
ListAttributes(first_number = 1L, style = "DefaultStyle",
  delim = "DefaultDelim")
}
\arguments{
\item{first_number}{an integer giving the first number of the list}

\item{style}{a character string giving the style, possible values are \code{"DefaultStyle"}, 
\code{"Example"}, \code{"Decimal"}, \code{"LowerRoman"}, 
\code{"UpperRoman"}, \code{"LowerAlpha"} and \code{"UpperAlpha"}.}

\item{delim}{a character string giving the delimiter, possible values are \code{"DefaultDelim"},
\code{"Period"}, \code{"OneParen"} and \code{"TwoParens"}.}
}
\description{
A constructor for pandoc list attributes.
}

