% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inline_elements.R
\name{Note}
\alias{Note}
\title{Note.}
\usage{
Note(x)
}
\arguments{
\item{x}{a pandoc block object or a list of pandoc block objects}
}
\description{
A constructor of an inline object of type \code{"Note"}.
}
\examples{
block <- Plain("x")
Note(block)

}
\seealso{
Other Inline element constructors: \code{\link{Cite}},
  \code{\link{Code}}, \code{\link{Emph}},
  \code{\link{Image}}, \code{\link{LineBreak}},
  \code{\link{Link}}, \code{\link{Math}},
  \code{\link{Quoted}}, \code{\link{RawInline}},
  \code{\link{SmallCaps}}, \code{\link{SoftBreak}},
  \code{\link{Space}}, \code{\link{Span}},
  \code{\link{Strikeout}}, \code{\link{Strong}},
  \code{\link{Str}}, \code{\link{Subscript}},
  \code{\link{Superscript}}
}
