% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/block_elements.R
\name{Null}
\alias{Null}
\title{Nothing.}
\usage{
Null()
}
\description{
Constructs a block object of type \code{Null}.
}
\examples{
Null()

}
\seealso{
Other Block element constructors: \code{\link{BlockQuote}},
  \code{\link{CodeBlock}}, \code{\link{DefinitionList}},
  \code{\link{Definition}}, \code{\link{Div}},
  \code{\link{Header}}, \code{\link{HorizontalRule}},
  \code{\link{OrderedList}}, \code{\link{Para}},
  \code{\link{Plain}}, \code{\link{Table}}
}
