% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{pandocfilters_writer}
\alias{pandocfilters_writer}
\title{Write the JSON-formatted AST to a connection}
\usage{
pandocfilters_writer(x, con, format)
}
\arguments{
\item{x}{a JSON representation of the AST to be written out}

\item{con}{a connection object or a character string to which the JSON-formatted AST is written}

\item{format}{a character string giving the format (e.g. \code{"latex"}, \code{"html"})}
}
\description{
Write the JSON-formatted AST to a connection.
}
\details{
If you want to apply a filter to the document before it get's written out, or your
         pandoc installation is not registered in the \code{PATH} it can be favorable to provide your
         own writer function to the document class.
}

