% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/block_elements.R
\name{Para}
\alias{Para}
\title{Paragraph.}
\usage{
Para(x)
}
\arguments{
\item{x}{a inline object or list of inline objects}
}
\description{
Constructs a block object of type \code{Para}.
}
\examples{
Para("x")

}
\seealso{
Other Block element constructors: \code{\link{BlockQuote}},
  \code{\link{CodeBlock}}, \code{\link{DefinitionList}},
  \code{\link{Definition}}, \code{\link{Div}},
  \code{\link{Header}}, \code{\link{HorizontalRule}},
  \code{\link{Null}}, \code{\link{OrderedList}},
  \code{\link{Plain}}, \code{\link{Table}}
}
