% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inline_elements.R
\name{RawInline}
\alias{RawInline}
\title{Raw Inline.}
\usage{
RawInline(format, x)
}
\arguments{
\item{format}{a character string giving the format (e.g. \code{"latex"}, \code{"html"})}

\item{x}{a character string giving the inline}
}
\description{
A constructor of an inline object of type \code{"RawInline"}.
}
\examples{
RawInline("latex", "some RawInline")

}
\seealso{
Other Inline element constructors: \code{\link{Cite}},
  \code{\link{Code}}, \code{\link{Emph}},
  \code{\link{Image}}, \code{\link{LineBreak}},
  \code{\link{Link}}, \code{\link{Math}},
  \code{\link{Note}}, \code{\link{Quoted}},
  \code{\link{SmallCaps}}, \code{\link{SoftBreak}},
  \code{\link{Space}}, \code{\link{Span}},
  \code{\link{Strikeout}}, \code{\link{Strong}},
  \code{\link{Str}}, \code{\link{Subscript}},
  \code{\link{Superscript}}
}
