% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{set_pandoc_path}
\alias{set_pandoc_path}
\title{Set Pandoc Path}
\usage{
set_pandoc_path(path = "pandoc")
}
\arguments{
\item{path}{a character giving the location of pandoc 
(default is \code{"pandoc"} which uses the pandoc set in the system path).}
}
\description{
Set the path to pandoc.
}
