% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{set_pandoc_version}
\alias{set_pandoc_version}
\title{Set Pandoc Version}
\usage{
set_pandoc_version(x)
}
\arguments{
\item{x}{a numeric giving the pandoc version (e.g. 1.14 or 1.15 or 1.16 or 1.17)}
}
\description{
Set the version version pandoc.
}

