% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inline_elements.R
\name{SoftBreak}
\alias{SoftBreak}
\title{Soft Line Break.}
\usage{
SoftBreak()
}
\description{
A constructor of an inline object of type \code{"SoftBreak"}.
}
\examples{
SoftBreak()

}
\seealso{
Other Inline element constructors: \code{\link{Cite}},
  \code{\link{Code}}, \code{\link{Emph}},
  \code{\link{Image}}, \code{\link{LineBreak}},
  \code{\link{Link}}, \code{\link{Math}},
  \code{\link{Note}}, \code{\link{Quoted}},
  \code{\link{RawInline}}, \code{\link{SmallCaps}},
  \code{\link{Space}}, \code{\link{Span}},
  \code{\link{Strikeout}}, \code{\link{Strong}},
  \code{\link{Str}}, \code{\link{Subscript}},
  \code{\link{Superscript}}
}
