% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inline_elements.R
\name{Space}
\alias{Space}
\title{Inter-word space.}
\usage{
Space()
}
\description{
A constructor of an inline object of type \code{"Space"}.
}
\examples{
Space()

}
\seealso{
Other Inline element constructors: \code{\link{Cite}},
  \code{\link{Code}}, \code{\link{Emph}},
  \code{\link{Image}}, \code{\link{LineBreak}},
  \code{\link{Link}}, \code{\link{Math}},
  \code{\link{Note}}, \code{\link{Quoted}},
  \code{\link{RawInline}}, \code{\link{SmallCaps}},
  \code{\link{SoftBreak}}, \code{\link{Span}},
  \code{\link{Strikeout}}, \code{\link{Strong}},
  \code{\link{Str}}, \code{\link{Subscript}},
  \code{\link{Superscript}}
}
