% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inline_elements.R
\name{Span}
\alias{Span}
\title{Generic Inline Container with Attributes.}
\usage{
Span(attr, inline)
}
\arguments{
\item{attr}{an object of type \link{Attr}}

\item{inline}{a inline object or a list of inline objects which will be shown}
}
\description{
A constructor of an inline object of type \code{"Span"}.
}
\examples{
attr <- Attr("A", "B", list(c("C", "D")))
Span(attr, "some inline string")

}
\seealso{
Other Inline element constructors: \code{\link{Cite}},
  \code{\link{Code}}, \code{\link{Emph}},
  \code{\link{Image}}, \code{\link{LineBreak}},
  \code{\link{Link}}, \code{\link{Math}},
  \code{\link{Note}}, \code{\link{Quoted}},
  \code{\link{RawInline}}, \code{\link{SmallCaps}},
  \code{\link{SoftBreak}}, \code{\link{Space}},
  \code{\link{Strikeout}}, \code{\link{Strong}},
  \code{\link{Str}}, \code{\link{Subscript}},
  \code{\link{Superscript}}
}
