% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pandocfilters.R
\name{Subscript}
\alias{Subscript}
\title{Subscripted Text}
\usage{
Subscript(x)
}
\arguments{
\item{x}{a inline object or a list of inline objects}
}
\description{
A constructor of an inline object of type \code{"Subscript"}.
}
\examples{
Subscript("some text written in superscript")
}

