% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inline_elements.R
\name{Superscript}
\alias{Superscript}
\title{Superscripted Text.}
\usage{
Superscript(x)
}
\arguments{
\item{x}{a inline object or a list of inline objects}
}
\description{
A constructor of an inline object of type \code{"Superscript"}.
}
\examples{
Superscript("some text written in superscript")

}
\seealso{
Other Inline element constructors: \code{\link{Cite}},
  \code{\link{Code}}, \code{\link{Emph}},
  \code{\link{Image}}, \code{\link{LineBreak}},
  \code{\link{Link}}, \code{\link{Math}},
  \code{\link{Note}}, \code{\link{Quoted}},
  \code{\link{RawInline}}, \code{\link{SmallCaps}},
  \code{\link{SoftBreak}}, \code{\link{Space}},
  \code{\link{Span}}, \code{\link{Strikeout}},
  \code{\link{Strong}}, \code{\link{Str}},
  \code{\link{Subscript}}
}
