% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pandocfilters.R
\name{TableCell}
\alias{TableCell}
\title{Table Cell}
\usage{
TableCell(x)
}
\arguments{
\item{x}{a character string giving the content of the table cell}
}
\description{
Table cells is a constructor for plain table cells.
}
\details{
In general table cells are a list of block elements, the 
         constructor \code{TableCell} creates a plain table cell.
}
\examples{
TableCell("Cell 1")
}

