% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{write.pandoc}
\alias{write.pandoc}
\title{Write the JSON-formatted AST to a connection}
\usage{
write.pandoc(json, file, format, exchange = c("arg", "file"))
}
\arguments{
\item{json}{a JSON representation of the AST to be written out}

\item{file}{a connection object or a character string to which the JSON-formatted AST is written}

\item{format}{a character string giving the format (e.g. \code{"latex"}, \code{"html"})}

\item{exchange}{a character string}
}
\description{
Write the JSON-formatted AST to a connection.
}
\details{
If you want to apply a filter to the document before it get's written out, or your
         pandoc installation is not registered in the \code{PATH} it can be favorable to provide your
         own writer function to the document class.
}
