% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pandocfilters.R
\name{BlockQuote}
\alias{BlockQuote}
\title{Block Quote}
\usage{
BlockQuote(blocks)
}
\arguments{
\item{blocks}{a block object or list of block objects}
}
\description{
A constructor of a block object of type \code{"BlockQuote"}.
}
\examples{
BlockQuote(Plain("Hello R!"))
}

