% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pandocfilters.R
\name{Cite}
\alias{Cite}
\title{Citation}
\usage{
Cite(citation, x)
}
\arguments{
\item{citation}{an object of type \code{"Citation"}}

\item{x}{a inline object or a list of inline objects}
}
\description{
A constructor of an inline object of type \code{"Cite"}.
}
\examples{
ci <- Citation(suffix=list(Str("Suffix_1")),
               id="Citation_ID_1", prefix=list(Str("Prefix_1")))
Cite(ci, Str("some text"))
}

