% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pandocfilters.R
\name{Code}
\alias{Code}
\title{Inline Code}
\usage{
Code(code, name = "", language = NULL, line_numbers = FALSE,
  start_from = 1)
}
\arguments{
\item{code}{a character string giving the inline code}

\item{name}{an optional character string giving the name of the inline code chunk}

\item{language}{an optional character string giving the programming language}

\item{line_numbers}{a logical which controls if line numbers should be used}

\item{start_from}{an integer giving the first line number}
}
\description{
A constructor of an inline object of type \code{"Code"}.
}
\examples{
Code("lm(hello ~ world)", "my_r_inline_code", "R", TRUE, 0)
Code("lm(hello ~ world)")
}

