% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pandocfilters.R
\name{CodeBlock}
\alias{CodeBlock}
\title{Code Block}
\usage{
CodeBlock(attr, code)
}
\arguments{
\item{attr}{an object of type \code{"Attr"}}

\item{code}{a character string containing the source code.}
}
\description{
A constructor of a block object of type \code{"CodeBlock"}.
}
\examples{
attr <- Attr("id", "Programming Language", list(c("key", "value")))
code <- "x <- 3\\nprint('Hello R!')"
CodeBlock(attr, code)
}

