% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pandocfilters.R
\name{Definition}
\alias{Definition}
\title{Definition}
\usage{
Definition(key, value)
}
\arguments{
\item{key}{a inline object or list of inline objects}

\item{value}{a block object or list of block objects}
}
\description{
A constructor of a \code{Definition} which can be used as an element of a \code{"DefinitionList"}.
}
\examples{
Definition("some key", Plain("some value"))
}

