% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pandocfilters.R
\name{Div}
\alias{Div}
\title{Generic Block Container with Attributes}
\usage{
Div(blocks, attr = Attr())
}
\arguments{
\item{blocks}{a block object or list of block objects}

\item{attr}{an object of type \code{"Attr"}}
}
\description{
A constructor of a block object of type \code{"Div"}.
}
\examples{
blocks <- Plain("Hello R!")
Div(blocks)
}

