% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pan_document.R
\docType{class}
\name{document}
\alias{document}
\title{Create a new Document.}
\usage{
document()
}
\description{
Class for creating documents using pandoc format.
}
\section{Usage}{

\preformatted{
d <- document()
d$append(x) 
d$append_block_quote(blocks) 
d$append_bullet_list(lblocks) 
d$append_code_block(attr, code) 
d$append_definition_list(x) 
d$append_div(blocks, attr) 
d$append_header(x, level = 1L, attr = Attr()) 
d$append_horizontal_rule() 
d$append_null() 
d$append_ordered_list(lattr, lblocks) 
d$append_para(x) 
d$append_plain(x) 
d$append_table(rows, col_names = NULL, aligns = NULL, col_width = NULL)
d$to_json() 
d$write(con, format = "markdown", writer = pandocfilters_writer)

print(d)
}
}

\section{Arguments}{

\describe{
\item{x}{a block object or list of block objects}
\item{blocks}{passed to \code{\link[=BlockQuote]{BlockQuote()}}}
\item{lbocks}{see \code{\link[=OrderedList]{OrderedList()}}}
\item{attr}{see \code{\link[=Attr]{Attr()}}}
\item{code}{see \code{\link[=CodeBlock]{CodeBlock()}}}
\item{level}{see \code{\link[=Header]{Header()}}}
\item{lattr}{see \code{\link[=OrderedList]{OrderedList()}}}
\item{lblocks}{see \code{\link[=BulletList]{BulletList()}}}
\item{rows}{see \code{\link[=Table]{Table()}}}
\item{col_names}{see \code{\link[=Table]{Table()}}}
\item{aligns}{see \code{\link[=Table]{Table()}}}
\item{col_width}{see \code{\link[=Table]{Table()}}}
\item{con}{a connection object or a character string to which the document is written}
\item{format}{a character string giving the format (e.g. \code{latex}, \code{html})}
\item{writer}{an optional writer function, see \code{\link[=pandocfilters_writer]{pandocfilters_writer()}}. Any function with the three arguments \code{x}, \code{con} and \code{format} can be used as writer function}
}
}

\section{Details}{

\code{append(x)} appends a new block to the document.

\code{append_plain(x)}: appends \code{\link[=Plain]{Plain()}}

\code{append_para(x)}: appends \code{\link[=Para]{Para()}}

\code{append_code_block()}:  appends \code{\link[=CodeBlock]{CodeBlock()}}

\code{append_block_quote()}: appends \code{\link[=BlockQuote]{BlockQuote()}}

\code{append_ordered_list()}: appends \code{\link[=OrderedList]{OrderedList()}}

\code{append_bullet_list()}: appends \code{\link[=BulletList]{BulletList()}}

\code{append_definition_list()}: appends \code{\link[=DefinitionList]{DefinitionList()}}

\code{append_header()}: appends \code{\link[=Header]{Header()}}

\code{append_horizontal_rule()}: appends \code{\link[=HorizontalRule]{HorizontalRule()}}

\code{append_table()}: appends \code{\link[=Table]{Table()}}

\code{append_div()}: appends \code{\link[=Div]{Div()}}

\code{append_null()}: appends \code{\link[=Null]{Null()}}

\code{to_json()} returns the \code{JSON} representation of the document. See also \code{\link[=to_pandoc_json]{to_pandoc_json()}}

\code{write()} writes the JSON-formatted AST to a connection. See also \code{\link[=pandocfilters_writer]{pandocfilters_writer()}}
}

