% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pandocfilters.R
\name{Emph}
\alias{Emph}
\title{Emphasized Text}
\usage{
Emph(x)
}
\arguments{
\item{x}{a inline object or a list of inline objects}
}
\description{
A constructor of an inline object of type \code{"Emph"}.
}
\examples{
Emph("emphasize")
}

