% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pandocfilters.R
\name{Header}
\alias{Header}
\title{Header}
\usage{
Header(x, level = 1L, attr = Attr())
}
\arguments{
\item{x}{a inline object or a list of inline objects}

\item{level}{an integer giving the level}

\item{attr}{an object of type \code{"Attr"}}
}
\description{
A constructor of a block object of type \code{"Header"}.
}
\examples{
Header("My Header")
}

