% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pandocfilters.R
\name{HorizontalRule}
\alias{HorizontalRule}
\title{Horizontal Rule}
\usage{
HorizontalRule()
}
\description{
A constructor of a block object of type \code{"HorizontalRule"}.
}
\examples{
HorizontalRule()
}

