% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pandocfilters.R
\name{Image}
\alias{Image}
\title{Image}
\usage{
Image(target, text, caption = "", attr = Attr())
}
\arguments{
\item{target}{a character string giving the target (hyper reference)}

\item{text}{a inline object or a list of inline objects giving the visible part}

\item{caption}{a character string describing the picture}

\item{attr}{an optional object of type \code{"Attr"}}
}
\description{
A constructor of an inline object of type \code{"Image"}.
}
\details{
Further Usage examples can be found in the README.
}
\examples{
Image("https:://Rlogo.jpg", "some_text", "fig:some_caption")
}

