% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pandocfilters.R
\name{Note}
\alias{Note}
\title{Note}
\usage{
Note(x)
}
\arguments{
\item{x}{a pandoc block object or a list of pandoc block objects}
}
\description{
A constructor of an inline object of type \code{"Note"}.
}
\examples{
block <- Plain("x")
Note(block)
}

