% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pandocfilters.R
\name{Para}
\alias{Para}
\title{Paragraph}
\usage{
Para(x)
}
\arguments{
\item{x}{a inline object or list of inline objects}
}
\description{
A constructor of a block object of type \code{"Para"}.
}
\examples{
Para("x")
}

