% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pandocfilters.R
\name{Plain}
\alias{Plain}
\title{Plain Text}
\usage{
Plain(x)
}
\arguments{
\item{x}{a inline object or list of inline objects}
}
\description{
Constructs a block object of type \code{"Plain"}, a plain paragraph.
}
\examples{
Plain("x")
}
