% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pandocfilters.R
\name{RawInline}
\alias{RawInline}
\title{Raw Inline}
\usage{
RawInline(format, x)
}
\arguments{
\item{format}{a character string giving the format (e.g. \code{"latex"}, \code{"html"})}

\item{x}{a character string giving the inline}
}
\description{
A constructor of an inline object of type \code{"RawInline"}.
}
\examples{
RawInline("latex", "some RawInline")
}

