% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pandocfilters.R
\name{SmallCaps}
\alias{SmallCaps}
\title{Small Caps Text}
\usage{
SmallCaps(x)
}
\arguments{
\item{x}{a inline object or a list of inline objects}
}
\description{
A constructor of an inline object of type \code{"SmallCaps"}.
}
\examples{
SmallCaps("The latex command for 'small caps' is 'textsc'!")
}

