% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pandocfilters.R
\name{Span}
\alias{Span}
\title{Generic Inline Container with Attributes}
\usage{
Span(attr, inline)
}
\arguments{
\item{attr}{an object of type \code{"Attr"}}

\item{inline}{a inline object or a list of inline objects which will be shown}
}
\description{
A constructor of an inline object of type \code{"Span"}.
}
\examples{
attr <- Attr("A", "B", list(c("C", "D")))
Span(attr, "some inline string")
}

