% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pandocfilters.R
\name{Strikeout}
\alias{Strikeout}
\title{Strikeout Text}
\usage{
Strikeout(x)
}
\arguments{
\item{x}{a inline object or a list of inline objects}
}
\description{
Constructs an inline object of type \code{"Strikeout"}.
}
\examples{
Strikeout("strikeout")
}
