% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pandocfilters.R
\name{Strong}
\alias{Strong}
\title{Strongly Emphasized Text}
\usage{
Strong(x)
}
\arguments{
\item{x}{a inline object or a list of inline objects}
}
\description{
A constructor of an inline object of type \code{"Strong"}.
}
\examples{
Strong("strong")
}

