% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pandocfilters.R
\name{Superscript}
\alias{Superscript}
\title{Superscripted Text}
\usage{
Superscript(x)
}
\arguments{
\item{x}{a inline object or a list of inline objects}
}
\description{
A constructor of an inline object of type \code{"Superscript"}.
}
\examples{
Superscript("some text written in superscript")
}

