% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pandocfilters.R
\name{Table}
\alias{Table}
\title{Table}
\usage{
Table(rows, col_names = NULL, aligns = NULL, col_width = NULL,
  caption = list())
}
\arguments{
\item{rows}{an object of class \code{"matrix"}, \code{"data.frame"}, \code{"table"} 
or a list of lists of pandoc objects of type \code{"TableCell"}}

\item{col_names}{a list of objects of type \code{"TableCell"}}

\item{aligns}{a character vector of alignments, possible values are \dQuote{l} for left,
\dQuote{r} for right, \dQuote{c} for center and \dQuote{d} for default.}

\item{col_width}{a numeric vector}

\item{caption}{a inline object or a list of inline objects giving the caption}
}
\description{
A constructor of a block object of type \code{"Table"}.
}
\details{
Table, with caption, column alignments (required), relative column widths 
         (0 = default), column headers (each a list of blocks), 
         and rows (each a list of lists of blocks)
}

