% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{pandoc_to_json}
\alias{pandoc_to_json}
\alias{pandoc_from_json}
\title{Utility functions for testing filters}
\usage{
pandoc_to_json(file, from = "markdown")

pandoc_from_json(json, to = "markdown", exchange = c("file", "arg"))
}
\arguments{
\item{file}{file name}

\item{from}{markdown, html, latex or native}

\item{json}{a \code{JSON} representation of the \code{AST} to be passed 
to \code{pandoc}}

\item{to}{markdown, html, latex or native}

\item{exchange}{a character string}
}
\description{
Utility functions for testing filters
}
